% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{addGaps}
\alias{addGaps}
\title{\code{addGaps} Add IMGT gaps to IgBlast output}
\usage{
addGaps(db, gapdb, organism = "human", locus = "Ig", gapped_d = FALSE)
}
\arguments{
\item{db}{AIRR-rearrangment formatted dataframe outputted from \link{assignGenes}}

\item{gapdb}{Root directory of gapped fasta files (\code{dir} in \link{downloadIMGT})}

\item{organism}{Organism (human, mouse, rhesus_monkey)}

\item{locus}{Ig or TR}

\item{gapped_d}{Include IMGT gaps in D regions? Only applicable to}
}
\value{
AIRR-rearrangement formatted data frame in which sequence_alignment and germline_alignment
have been updated with IMGT gaps
}
\description{
\code{addGaps} Add IMGT gaps to IgBlast output
}
\details{
Similar functionality to MakeDb.py in Change-O.
}
