% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{assignGenes}
\alias{assignGenes}
\title{\code{assignGenes} Runs IgBlast on a fasta file}
\usage{
assignGenes(
  file,
  igblast,
  refs,
  igdata = NULL,
  organism = "human",
  domain_system = "imgt",
  outfile = NULL,
  nproc = 1,
  db_prefix = "imgt",
  locus = "Ig",
  set_igdata = TRUE,
  return = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{file}{Fasta file of Ig or TR sequences}

\item{igblast}{Location of IgBlast program directory, containing /bin/igblastn}

\item{refs}{Reference directory of sequences (see downloadIMGT)}

\item{igdata}{Internal data directory for IgBlast (defaults to \code{igblast}/internal_data)}

\item{organism}{Organism (human, mouse, rhesus_monkey)}

\item{domain_system}{Currently only IMGT supported}

\item{outfile}{Name of AIRR rearrangement file (must end in TSV)}

\item{nproc}{Nummber of threads to use}

\item{db_prefix}{File prefix for reference fastas}

\item{locus}{Ig or TR}

\item{set_igdata}{Set IGDATA environment variable?}

\item{return}{Return data.frame of output?}

\item{verbose}{Print extra info?}
}
\value{
AIRR-rearrangement formatted data frame
}
\description{
\code{assignGenes} Runs IgBlast on a fasta file
}
\details{
Runs IgBlast, similar to AssignGenes.py in Changeo.
Must have IgBlast downloaded, precompiled binaries recommended:
https://ncbi.github.io/igblast/

Note for M1/M2 Mac OS, may be necessary to install IgBlast .dmg and 
manually set igdata. Otherwise most issues stem from the internal_data
directory location.
}
