% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{buildAllClonalGermlines}
\alias{buildAllClonalGermlines}
\title{\link{buildAllClonalGermlines} Determines and builds all possible germlines for a clone}
\usage{
buildAllClonalGermlines(
  receptors,
  references,
  chain = "IGH",
  use_regions = FALSE,
  vonly = FALSE,
  seq = "sequence_alignment",
  id = "sequence_id",
  clone = "clone_id",
  v_call = "v_call",
  j_call = "j_call",
  v_germ_start = "v_germline_start",
  v_germ_end = "v_germline_end",
  v_germ_length = "v_germline_length",
  d_germ_start = "d_germline_start",
  d_germ_end = "d_germline_end",
  d_germ_length = "d_germline_length",
  j_germ_start = "j_germline_start",
  j_germ_end = "j_germline_end",
  j_germ_length = "j_germline_length",
  np1_length = "np1_length",
  np2_length = "np2_length",
  j_germ_aa_length = "j_germline_aa_length",
  amino_acid = FALSE,
  threshold = 3,
  ...
)
}
\arguments{
\item{receptors}{AIRR-table containing sequences from one clone}

\item{references}{Full list of reference segments, see \link{readIMGT}}

\item{chain}{chain in \code{references} being analyzed}

\item{use_regions}{Return string of VDJ regions? (optional)}

\item{vonly}{Return germline of only v segment?}

\item{seq}{Column name for sequence alignment}

\item{id}{Column name for sequence ID}

\item{clone}{Column name for clone ID}

\item{v_call}{Column name for V gene segment gene call}

\item{j_call}{Column name for J gene segment gene call}

\item{v_germ_start}{Column name of index of V segment start within germline}

\item{v_germ_end}{Column name of index of V segment end within germline}

\item{v_germ_length}{Column name of index of V segment length within germline}

\item{d_germ_start}{Column name of index of D segment start within germline}

\item{d_germ_end}{Column name of index of D segment end within germline}

\item{d_germ_length}{Column name of index of D segment length within germline}

\item{j_germ_start}{Column name of index of J segment start within germline}

\item{j_germ_end}{Column name of index of J segment end within germline}

\item{j_germ_length}{Column name of index of J segment length within germline}

\item{np1_length}{Column name in receptor specifying np1 segment length}

\item{np2_length}{Column name in receptor specifying np2 segment length}

\item{j_germ_aa_length}{Column name of J segment amino acid length (if amino_acid=TRUE)}

\item{amino_acid}{Perform reconstruction on amino acid sequence (experimental)}

\item{threshold}{The maximum germline length difference within a clone. Default is 3.}

\item{...}{Additional arguments passed to \link{buildGermline}}
}
\value{
A data frame with all possible reconstructed germlines
}
\description{
\link{buildAllClonalGermlines} Determines and builds all possible germlines for a clone
}
\seealso{
\link{createAllGermlines} \link{buildGermline}, \link{stitchVDJ}
}
