% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{buildBeast2Strict}
\alias{buildBeast2Strict}
\title{Writes a BEAST XML file given a list of airrClone objects, then runs them}
\usage{
buildBeast2Strict(
  data,
  exec,
  time,
  mcmc_length = 1e+06,
  dir = ".",
  include_gm = FALSE,
  nproc = 1,
  log_every = NULL,
  verbose = 1,
  burnin = 10,
  tree_prior = c("coalescent-skyline", "coalescent-constant"),
  pop_sizes = 5,
  tree_posterior = FALSE
)
}
\arguments{
\item{data}{an \code{airrClone} object}

\item{time}{name of time column (trait in getTrees)}

\item{mcmc_length}{Length of MCMC chain}

\item{dir}{directory where temporary files will be placed}

\item{include_gm}{Include germline sequence?}

\item{log_every}{log every X samples, default mcmc_length/10000}

\item{verbose}{if > 0 print run information}

\item{burnin}{percent of initial tree samples to discard (1-100)}

\item{tree_prior}{prior for tree model}

\item{pop_sizes}{number of population sizes for coalescent-skyline prior}

\item{tree_posterior}{store tree posterior? Used for densitree plots}

\item{germline}{name of germline sequence}

\item{units}{time units}
}
\value{
The input clones tibble with an additional column for the bootstrap replicate trees.
}
\description{
\code{buildBeast2Strict} Phylogenetic bootstrap function.
}
