% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{buildIgphyml}
\alias{buildIgphyml}
\title{Wrapper to build IgPhyML trees and infer intermediate nodes}
\usage{
buildIgphyml(
  clone,
  igphyml,
  trees = NULL,
  nproc = 1,
  temp_path = NULL,
  id = NULL,
  rseed = NULL,
  quiet = 0,
  rm_files = TRUE,
  rm_dir = NULL,
  ...
)
}
\arguments{
\item{clone}{\code{airrClone} object}

\item{igphyml}{igphyml executable}

\item{trees}{list of tree topologies if desired}

\item{nproc}{number of cores for parallelization}

\item{temp_path}{path to temporary directory}

\item{id}{IgPhyML run id}

\item{rseed}{random number seed if desired}

\item{quiet}{amount of rubbish to print}

\item{rm_files}{remove temporary files?}

\item{rm_dir}{remove temporary directory?}

\item{...}{Additional arguments (not currently used)}
}
\value{
\code{phylo} object created by igphyml with nodes attribute
         containing reconstructed sequences.
}
\description{
Wrapper to build IgPhyML trees and infer intermediate nodes
}
