% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{buildPhylo}
\alias{buildPhylo}
\title{Wrapper for alakazam::buildPhylipLineage}
\usage{
buildPhylo(
  clone,
  exec,
  temp_path = NULL,
  verbose = 0,
  rm_temp = TRUE,
  seq = "sequence",
  tree = NULL,
  onetree = TRUE
)
}
\arguments{
\item{clone}{\code{airrClone} object}

\item{exec}{dnapars or dnaml executable}

\item{temp_path}{path to temporary directory}

\item{verbose}{amount of rubbish to print}

\item{rm_temp}{remove temporary files?}

\item{seq}{sequece column in \code{airrClone} object}

\item{tree}{fixed tree topology if desired (currently does nothing
if specified)}

\item{onetree}{Only sample one tree if multiple found.}
}
\value{
\code{phylo} object created by dnapars or dnaml with nodes attribute
         containing reconstructed sequences.
}
\description{
Wrapper for alakazam::buildPhylipLineage
}
