% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{buildPML}
\alias{buildPML}
\title{Wrapper for phangorn::optim.pml}
\usage{
buildPML(
  clone,
  seq = "sequence",
  model = "GTR",
  gamma = FALSE,
  asr = "seq",
  asr_thresh = 0.05,
  tree = NULL,
  ...
)
}
\arguments{
\item{clone}{\code{airrClone} object}

\item{seq}{sequece column in \code{airrClone} object}

\item{model}{substitution model to use}

\item{gamma}{gamma site rate variation?}

\item{asr}{return sequence or probability matrix?}

\item{asr_thresh}{threshold for including a nucleotide as an alternative}

\item{tree}{fixed tree topology if desired.}

\item{...}{Additional arguments (not currently used)}
}
\value{
\code{phylo} object created by phangorn::optim.pml with nodes
         attribute containing reconstructed sequences.
}
\description{
Wrapper for phangorn::optim.pml
}
