% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{buildPratchet}
\alias{buildPratchet}
\title{Wrapper for phangorn::pratchet}
\usage{
buildPratchet(
  clone,
  seq = "sequence",
  asr = "seq",
  asr_thresh = 0.05,
  tree = NULL,
  asr_type = "MPR",
  verbose = 0,
  resolve_random = TRUE,
  data_type = "DNA",
  ...
)
}
\arguments{
\item{clone}{\code{airrClone} object}

\item{seq}{sequece column in \code{airrClone} object}

\item{asr}{return sequence or probability matrix?}

\item{asr_thresh}{threshold for including a nucleotide as an alternative}

\item{tree}{fixed tree topology if desired.}

\item{asr_type}{MPR or ACCTRAN}

\item{verbose}{amount of rubbish to print}

\item{resolve_random}{randomly resolve polytomies?}

\item{data_type}{Are sequences DNA or AA?}

\item{...}{Additional arguments (not currently used)}
}
\value{
\code{phylo} object created by phangorn::pratchetet with nodes
         attribute containing reconstructed sequences.
}
\description{
Wrapper for phangorn::pratchet
}
