% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{checkNodeDivergences}
\alias{checkNodeDivergences}
\title{For each node in nodes list, check that the current divergence
is the same as previously reported}
\usage{
checkNodeDivergences(tree, stop = TRUE, catch_null = TRUE)
}
\arguments{
\item{tree}{A \code{phylo} object with nodes list}

\item{stop}{Throw an error if divergences not equal}

\item{catch_null}{Should null divergence values count as mismatches?}
}
\value{
TRUE or FALSE depening on whether node divergences are consistent
if stop=TRUE, a FALSE value will cause an error
}
\description{
For each node in nodes list, check that the current divergence
is the same as previously reported
}
