% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{colorTrees}
\alias{colorTrees}
\title{Get a color palette for a predefined set of trait values}
\usage{
colorTrees(trees, palette, ambig = "blend")

colorTrees(trees, palette, ambig = "blend")
}
\arguments{
\item{trees}{list of phylo objects with assigned internal node states}

\item{palette}{named vector of colors (see \link{getPalette})}

\item{ambig}{how should ambiguous states be colored (blend or grey)}
}
\value{
A list of colored trees

A list of colored trees
}
\description{
\code{colorTree} Gets a color palette for a predefined set of trait values

\code{colorTree} Gets a color palette for a predefined set of trait values
}
\details{
Trees must have node states represented in a "states" vector. By default,
ambiguous states (separated by ",") have their colors blended. If

Trees must have node states represented in a "states" vector. By default,
ambiguous states (separated by ",") have their colors blended. If
}
\seealso{
\link{getPalette}, \link{getTrees}, \link{plotTrees}

\link{getPalette}, \link{getTrees}, \link{plotTrees}
}
