% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{condenseTrees}
\alias{condenseTrees}
\title{Condense a set of equally parsimonious node labels into a single tree}
\usage{
condenseTrees(trees, states, palette)
}
\arguments{
\item{trees}{List of the same tree with equally parsimonious labels}

\item{states}{States in model}

\item{palette}{Named vector with a color per state}
}
\value{
Unhelpful
}
\description{
\code{condenseTrees} Condenses a set of equally parsimonious node labels into a single tree
}
\details{
None right now.
}
\examples{
# Example data

}
\seealso{
Filler
}
