% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTreesFunctions.R
\name{create_alignment}
\alias{create_alignment}
\title{Takes an airr clone object and returns BEAST2 Alignment xml of the sequences}
\usage{
create_alignment(clone, id, include_germline_as_tip)
}
\arguments{
\item{clone}{an \code{airrClone} object}

\item{id}{unique identifer for this analysis}

\item{include_germline_as_tip}{include the germline as a tip in the alignment?}
}
\value{
String of BEAST2 Alignment and TaxonSet xml
}
\description{
Takes an airr clone object and returns BEAST2 Alignment xml of the sequences
}
