% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTreesFunctions.R
\name{create_height_prior}
\alias{create_height_prior}
\title{Takes an airr clone object and returns BEAST2 XML to set a height prior}
\usage{
create_height_prior(clone, id, start_date)
}
\arguments{
\item{clone}{an \code{airrClone} object}

\item{id}{unique identifer for this analysis}

\item{start_date}{starting date to use as prior, in forward time}
}
\value{
String of XML setting the height prior
}
\description{
Takes an airr clone object and returns BEAST2 XML to set a height prior
}
