% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTreesFunctions.R
\name{create_max_height_prior}
\alias{create_max_height_prior}
\title{Takes an airr clone object and returns BEAST2 XML to set a maximum height prior}
\usage{
create_max_height_prior(clone, id, max_start_date)
}
\arguments{
\item{clone}{an \code{airrClone} object}

\item{id}{unique identifer for this analysis}

\item{max_start_date}{max start date to use for prior, in forward time}
}
\value{
String of XML setting the MRCA prior of the observed sequences
}
\description{
Takes an airr clone object and returns BEAST2 XML to set a maximum height prior
}
