% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTreesFunctions.R
\name{create_MRCA_prior_germline}
\alias{create_MRCA_prior_germline}
\title{Takes an airr clone object and returns BEAST2 XML for MRCA prior of the germline sequence}
\usage{
create_MRCA_prior_germline(clone, id, germline_range)
}
\arguments{
\item{clone}{an \code{airrClone} object}

\item{id}{unique identifer for this analysis}

\item{germline_range}{Possible date range of germline tip}
}
\value{
String of XML setting the MRCA prior of the germline sequence
}
\description{
Takes an airr clone object and returns BEAST2 XML for MRCA prior of the germline sequence
}
