% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTreesFunctions.R
\name{create_MRCA_prior_observed}
\alias{create_MRCA_prior_observed}
\title{Takes an airr clone object and returns BEAST2 XML for MRCA prior of the observed sequences}
\usage{
create_MRCA_prior_observed(clone, id)
}
\arguments{
\item{clone}{an \code{airrClone} object}

\item{id}{unique identifer for this analysis}
}
\value{
String of XML setting the MRCA prior of the observed sequences
}
\description{
Takes an airr clone object and returns BEAST2 XML for MRCA prior of the observed sequences
}
