% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTreesFunctions.R
\name{create_root_freqs}
\alias{create_root_freqs}
\title{Takes an airr clone object and returns BEAST2 rootfreqs xml of the germline}
\usage{
create_root_freqs(clone, id)
}
\arguments{
\item{clone}{an \code{airrClone} object}

\item{id}{unique identifer for this analysis}
}
\value{
String of XML setting the root frequencies to the germline sequence
}
\description{
Takes an airr clone object and returns BEAST2 rootfreqs xml of the germline
}
