% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{downloadIMGT}
\alias{downloadIMGT}
\title{\code{downloadIMGT} Download IMGT GENE-DB databases}
\usage{
downloadIMGT(
  dir = "imgt",
  organism = c("human", "mouse", "rat", "rabbit", "rhesus_monkey"),
  timeout = 300
)
}
\arguments{
\item{dir}{directory to contain database}

\item{organism}{vector of species to download: human, mouse, rat, rabbit, rhesus_monkey}

\item{timeout}{max time allowed for each download (temporarily sets global timeout option)}
}
\value{
directory of downloaded IMGT BCR and TCR reference files
}
\description{
Loads all reference germlines from an Immcantation-formatted IMGT database.
}
\details{
Recoding of this script: 
https://bitbucket.org/kleinstein/immcantation/src/master/scripts/fetch_imgtdb.sh
}
