% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{downsampleClone}
\alias{downsampleClone}
\title{\code{downsampleClone} Down-sample clone to maximum tip/switch ratio}
\usage{
downsampleClone(clone, trait, tip_switch = 20, tree = NULL)
}
\arguments{
\item{clone}{an \link{airrClone} object}

\item{trait}{trait considered for rarefaction
\link{getTrees}}

\item{tip_switch}{maximum tip/switch ratio}

\item{tree}{a \code{phylo} tree object correspond to \code{clone}}
}
\value{
A vector with sequence for each locus at a specified \code{node}
          in \code{tree}.
}
\description{
\code{downsampleClone} Down-sample clone to maximum tip/switch ratio
}
