% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clones.R
\name{dowserObjectEqual}
\alias{dowserObjectEqual}
\title{\code{dowserObjectsEqual}}
\usage{
dowserObjectEqual(obj1, obj2, verbose = TRUE, edge_tol = 1e-08)
}
\arguments{
\item{obj1}{First Dowser object}

\item{obj2}{Second Dowser object}

\item{verbose}{Print number of checks performed}

\item{edge_tol}{Allowable difference among edge lengths}
}
\value{
Error if objects not equivalent, 0 all pass
}
\description{
\code{dowserObjectsEqual}
}
