% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ExampleAirrTyCHE}
\alias{ExampleAirrTyCHE}
\title{Example AIRR database for TyCHE}
\format{
A data.frame with the following AIRR style columns:
  \itemize{
    \item  \code{sequence_id}:           Sequence identifier
    \item  \code{sequence_alignment}:    IMGT-gapped observed sequence.
    \item  \code{germline_alignment}:    IMGT-gapped germline sequence.
    \item  \code{v_call}:                V region allele assignments.
    \item  \code{d_call}:                D region allele assignments.
    \item  \code{j_call}:                J region allele assignments.
    \item  \code{junction}:              Junction region sequence.
    \item  \code{junction_length}:       Length of the junction region in nucleotides.
    \item  \code{np1_length}:            Combined length of the N and P regions proximal
                                         to the V region.
    \item  \code{np2_length}:            Combined length of the N and P regions proximal
                                         to the J region.
    \item  \code{sample_time}:           Time point of the sample.
    \item  \code{location}:              Location of tissue from which the sample was taken.
    \item  \code{clone_id}:              Clonal group identifier.
}
}
\usage{
ExampleAirrTyCHE
}
\description{
A small example database from simble.
}
\references{
\enumerate{
  \item  Pre-submission.
}
}
\keyword{datasets}
