% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ExampleClones}
\alias{ExampleClones}
\title{Example Ig lineage trees}
\format{
A tibble of airrClone and phylo objects output by getTrees.
  \itemize{
    \item  \code{clone_id}:   Clonal cluster
    \item  \code{data}:       List of airrClone objects
    \item  \code{seqs}:       Number of sequences
    \item  \code{trees}:      List of phylo objects
  }
}
\usage{
ExampleClones
}
\description{
A tibble of Ig lineage trees generated from the \code{ExampleAirr} file
}
\seealso{
\link{ExampleClones}
}
\keyword{datasets}
