% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ExampleMixedClones}
\alias{ExampleMixedClones}
\title{Example Multiple Partition Trees}
\format{
A data.frame with the following Change-O style columns:
  \itemize{
    \item  \code{clone_id}:           Clonal cluster
    \item  \code{data}:               List of airrClone objects
    \item  \code{locus}:              Locus identifier.
    \item  \code{seqs}:               Number of sequences
    \item \code{igphyml_partitioned_trees}:      IgPhyML partitioned tree
    \item \code{raxml_partitioned_trees}:        RAxML partitioned tree
}
}
\usage{
ExampleMixedClones
}
\description{
A small example database subset from Turner, J. S. et al. Human germinal centres 
engage memory and naive B cells after influenza vaccination. Nature 586, 127–132 (2020).
}
\keyword{datasets}
