% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clones.R
\name{filterPartialSeqs}
\alias{filterPartialSeqs}
\title{\code{filterPartialSeqs}}
\usage{
filterPartialSeqs(
  data,
  seq = "sequence_alignment",
  cutoff = 250,
  pattern = "[ATCG]",
  verbose = TRUE
)
}
\arguments{
\item{data}{Data table of sequences (preferably in AIRR format)}

\item{seq}{name of sequence column for filtering}

\item{cutoff}{number of matches to the \code{pattern} option required}

\item{pattern}{regex for matching (defaults to A, T, C, or G)}

\item{verbose}{print out number of seqs removed
data(ExampleAirr)
filtered = filterPartialSeqs(ExampleAirr)}
}
\description{
\code{filterPartialSeqs}
}
