% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{getAllSeqs}
\alias{getAllSeqs}
\title{Return all tip and internal node sequences}
\usage{
getAllSeqs(data, imgt_gaps = TRUE)
}
\arguments{
\item{data}{a tibble of \code{airrClone} objects with reconstructed trees, 
the output of \link{getTrees}}

\item{imgt_gaps}{include a column of gapped sequences?}
}
\value{
A tibble with sequence information for each tip and internal node
of a set of trees.
}
\description{
\code{getNodeSeq} Sequence retrieval function.
}
\details{
Column names:
clone_id = clone id 
node_id = name of node, either the sequence name if a tip or Node<number> if internal node
node = node number in tree. Tips are nodes 1:<number of tips>.
locus = locus of sequence
sequence = ungapped sequence, either observed for tips or reconstructed for internal nodes
sequence_alignment = sequence with IMGT gaps (optional)
}
\seealso{
\link{getTrees} \link{getNodeSeq}
}
