% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTreesFunctions.R
\name{getDiffPoint}
\alias{getDiffPoint}
\title{Recurse up to tree to find most recent node with different state, or the root}
\usage{
getDiffPoint(
  tree,
  targetnode,
  trait,
  height = "CAheight_mean",
  verbose = FALSE
)
}
\arguments{
\item{tree}{a treedata object, from getTimeTrees}

\item{targetnode}{current node}

\item{trait}{column name of the trait of interest}

\item{height}{which height value to return}

\item{verbose}{print out run info}
}
\description{
\code{getDiffPoint}
}
