% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{getDivergence}
\alias{getDivergence}
\title{Get divergence from MRCA for each tip}
\usage{
getDivergence(phy, minlength = 0.001)
}
\arguments{
\item{phy}{Tree object}

\item{minlength}{Branch lengths to collapse in trees}
}
\value{
A named vector of each tip's divergence from the tree's MRCA.
}
\description{
\code{rootToTop} performs root-to-tip regression permutation test
}
