% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTreesFunctions.R
\name{getHeightsAndLengths}
\alias{getHeightsAndLengths}
\title{Add height and length columns to a tree@data data frame}
\usage{
getHeightsAndLengths(tree)
}
\arguments{
\item{tree}{a treedata objects from read.beast}
}
\description{
Add height and length columns to a tree@data data frame
}
