% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{getPalette}
\alias{getPalette}
\title{Get a color palette for a predefined set of trait values}
\usage{
getPalette(states, palette, custom = FALSE)
}
\arguments{
\item{states}{states in model}

\item{palette}{The colorbrewer palette to use}

\item{custom}{Use a custom internal node palette?}
}
\value{
A named vector with each state corresponding to a color
}
\description{
\code{getPalette} Gets a color palette for a predefined set of trait values
}
\seealso{
\link{getTrees}, \link{plotTrees}
}
