% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{getParams}
\alias{getParams}
\title{Get parameter estimates from posterior distributions}
\usage{
getParams(clones, burnin = 10, tracefile = NULL, width = 8.5, height = 11, ...)
}
\arguments{
\item{clones}{A \code{tibble} object containing \code{airrClone} and 
parameters_posterior column}

\item{burnin}{percent of initial samples to discard (1-100)}

\item{tracefile}{file name for trace plots}

\item{width}{width of plot in inches}

\item{height}{height of plot in inches}

\item{...}{optional arguments passed to grDevices::pdf}
}
\value{
A \code{tibble} with the same columns as clones, but additional
column parameters corresponding to parameter estimates
}
\description{
\code{getParams} performs root-to-tip regression date randomization test
}
