% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{getSeq}
\alias{getSeq}
\title{Return IMGT gapped sequence of specified tree node}
\usage{
getSeq(node, data, tree = NULL, clone = NULL, gaps = TRUE)
}
\arguments{
\item{node}{numeric node in tree (see details)}

\item{data}{a tibble of \code{airrClone} objects, the output of 
\link{getTrees}}

\item{tree}{a \code{phylo} tree object containing \code{node}}

\item{clone}{if \code{tree} not specified, supply clone ID in \code{data}}

\item{gaps}{add IMGT gaps to output sequences?}
}
\value{
A vector with sequence for each locus at a specified \code{node}
          in \code{tree}.
}
\description{
\code{getSeq} Sequence retrieval function.
}
\details{
Use plotTrees(trees)[[1]] + geom_label(aes(label=node))+geom_tippoint() to show
node labels, and getSeq to return internal node sequences
}
\seealso{
\link{getTrees}
}
