% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{getSeqPath}
\alias{getSeqPath}
\title{Return all sequences along the tree from the germline to a specified node}
\usage{
getSeqPath(
  data,
  node,
  tree = NULL,
  clone = NULL,
  gaps = TRUE,
  translate = FALSE,
  rm_x_aa = FALSE,
  dot_notation = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{a tibble of \code{airrClone} objects, the output of 
\link{getTrees}}

\item{node}{numeric node in tree (see details)}

\item{tree}{a \code{phylo} tree object containing \code{node}}

\item{clone}{if \code{tree} not specified, supply clone ID in \code{data}}

\item{gaps}{add IMGT gaps to output sequences?}

\item{translate}{Tranlate into amino acids?}

\item{rm_x_aa}{Remove amino acids translated to X}

\item{dot_notation}{Represent sites with germline character as "." and ambiguous sites as "?"}

\item{verbose}{print out extra information?}
}
\value{
A data table where each row shows a node/node id, with that node's sequence(s)
in the remaining columns with respective locus labels.
}
\description{
Return all sequences along the tree from the germline to a specified node
}
\seealso{
\link{getNodeSeq} \link{getAllSeqs} \link{dfToFasta}
}
