% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTreesFunctions.R
\name{getSkylines}
\alias{getSkylines}
\title{Make data frames for Bayesian skyline plots}
\usage{
getSkylines(
  clones,
  time,
  bins = 100,
  verbose = 0,
  forward = TRUE,
  nproc = 1,
  max_height = c("min", "median", "mean", "max"),
  exclude_germline = TRUE
)
}
\arguments{
\item{clones}{clone tibble}

\item{time}{name of time column}

\item{bins}{number of bins for plotting}

\item{verbose}{if 1, print name of clones}

\item{forward}{plot in forward or (FALSE) backward time?}

\item{nproc}{processors for parallelization (by clone)}

\item{max_height}{max height to use (min, median, mean, max)}

\item{exclude_germline}{exclude germline from skyline plot? (For TyCHE GRTs)}
}
\value{
Bayesian Skyline values for given clone
}
\description{
\code{makeSkylines}
}
\details{
Clones must contain treedata objects with parameters_posterior and
trees_posterior. See \code{readBEAST} or \code{getTimeTrees} with posterior="all"
}
