% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{getTreesAndUCA}
\alias{getTreesAndUCA}
\title{\link{getTreesAndUCA} Construct trees and infer the UCA}
\usage{
getTreesAndUCA(
  clones,
  dir = NULL,
  build,
  exec,
  model_folder,
  uca_script,
  id = "sample",
  max_iters = 100,
  nproc = 1,
  rm_temp = TRUE,
  quiet = 0,
  omega = NULL,
  optimize = "lr",
  motifs = "FCH",
  hotness = "e,e,e,e,e,e",
  ...
)
}
\arguments{
\item{clones}{AIRR-table containing sequences \link{formatClones}}

\item{dir}{The file path of the directory of where data is saved. NULL is default.}

\item{build}{Name of the tree building method}

\item{exec}{File path to the tree building executable}

\item{model_folder}{The file path to the OLGA default model files}

\item{uca_script}{The file path to the UCA python script}

\item{id}{The run ID}

\item{max_iters}{The maximum number of iterations to run before ending}

\item{nproc}{The number of cores to use}

\item{rm_temp}{Remove the generated files?}

\item{quiet}{Amount of noise to print out}

\item{omega}{Omega parameters to estimate (see IgPhyMl docs)}

\item{optimize}{Optimize HLP rates (r), lengths (l), and/or topology (r)}

\item{motifs}{Motifs to consider (see IgPhyMl docs)}

\item{hotness}{Hotness parameters to estimate (see IgPhyMl docs)}

\item{...}{Additional arguments passed to \link{buildGermline}}
}
\value{
An \code{airrClone} object with trees and the inferred UCA
}
\description{
\link{getTreesAndUCA} Construct trees and infer the UCA
}
\details{
Return object adds/edits following columns:
\itemize{
  \item  \code{trees}:  The phylogenies associated with each clone
  \item  \code{UCA}:    The inferred UCA
}
}
\seealso{
\link{getTrees}
}
