% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{getTreesAndUCAs}
\alias{getTreesAndUCAs}
\title{\link{getTreesAndUCAs} Construct trees and infer the UCA}
\usage{
getTreesAndUCAs(
  clones,
  data,
  exec,
  model_folder,
  references,
  dir = NULL,
  model_folder_igk = NULL,
  model_folder_igl = NULL,
  partition = "single",
  repertoire_wide = FALSE,
  python = "python3",
  id = "sample",
  max_iters = 100,
  nproc = 1,
  rm_temp = TRUE,
  quiet = 0,
  chain = "H",
  clone = "clone_id",
  cell = "cell_id",
  subsample_size = NA,
  subsampling_method = c("random", "weighted", "least_mutated"),
  search = c("codon", "nt"),
  resolve_vj = FALSE,
  fix_vj_in_cdr3 = TRUE,
  fill_partials = TRUE,
  split_light = FALSE,
  ...
)
}
\arguments{
\item{clones}{AIRR-table containing sequences \link{formatClones}}

\item{data}{The AIRR-table that was used to make the clones object.}

\item{exec}{File path to the tree building executable}

\item{model_folder}{The file path to the OLGA default model files for heavy chains}

\item{references}{Reference genes. See \link{readIMGT}}

\item{dir}{The file path of the directory of where data is saved. NULL is default.}

\item{model_folder_igk}{The file path to the OLGA default model files for IGK}

\item{model_folder_igl}{The file path to the OLGA default model files for IGL}

\item{partition}{The partition model to use with IgPhyML. "single" is the default.}

\item{repertoire_wide}{Build trees using parameters inferred from the entire dataset?}

\item{python}{Specify the python call for your system. This is the call
on command line that issues the python you want to use. 
"python3" by default.}

\item{id}{The run ID, sample by default}

\item{max_iters}{The maximum number of iterations to run before ending. 
100 by default}

\item{nproc}{The number of cores to use}

\item{rm_temp}{Remove the generated files?}

\item{quiet}{Amount of noise to print out}

\item{chain}{Set to HL to use both heavy and light chain sequences}

\item{clone}{The name of the clone id column used in \link{formatClones}.}

\item{cell}{The name of the cell id in the AIRR table used to generate \link{formatClones}}

\item{subsample_size}{The amount that the clone should be sampled down to. 
By default this is NA to not induce subsampling.}

\item{subsampling_method}{How to subsample. Methods include 'random', 'weighted', 
and 'least_mutated'. The later two methods
require 'mu_freq' to be passed as a trait when running}

\item{search}{Search codon or nt space}

\item{resolve_vj}{Resolve the V and J gene annotations within each clone?}

\item{fix_vj_in_cdr3}{Check if the inferred V/J lengths go into the inferred cdr3 region and adjust accordingly.}

\item{fill_partials}{A logical that will fill in the V and J UCAs of clones that have partial V/J sequence alignments}

\item{split_light}{A logical that indicates if different light chain groups should be used to further split a clone (recommended for paired data)}

\item{...}{Additional arguments passed to various other functions like \link{getTrees} and \link{buildGermline}}
}
\value{
An \code{airrClone} object with trees and the inferred UCA
}
\description{
\link{getTreesAndUCAs} Construct trees and infer the UCA
}
\details{
Return object adds/edits following columns:
\itemize{
  \item  \code{trees}:  The phylogenies associated with each clone
  \item  \code{UCA}:    The inferred UCA
}
}
\seealso{
\link{getTrees}
}
