% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{getTreeSimulation}
\alias{getTreeSimulation}
\title{Simulates a dataset based on a given tree input within a \code{airrClone} objects.}
\usage{
getTreeSimulation(
  clones,
  naive_heavy_chain,
  naive_light_chain = NULL,
  targetingModel = "HH_S5F",
  lc_targetingModel = "HKL_S5F",
  nproc = 1,
  mutational_rate = 2,
  cell_id = "cell_id",
  sequence_alignment = "sequence_alignment",
  clone_id = "clone_id",
  sequence_id = "sequence_id",
  dir = NULL,
  id = NULL
)
}
\arguments{
\item{clones}{tibble \code{airrClone} objects, the output of 
\link{formatClones}}

\item{naive_heavy_chain}{The naive heavy chain object}

\item{naive_light_chain}{The naive light chain object}

\item{targetingModel}{The targeting model to use for the heavy chain simulations. 
Options include HH_S1F, HH_S5F, HKL_S1F, HKL_S5F, MK_RS1NF, 
MK_RS5NF, and U5N}

\item{lc_targetingModel}{The targeting model to use for the light chain simulations. 
Options include HH_S1F, HH_S5F, HKL_S1F, HKL_S5F, MK_RS1NF, 
MK_RS5NF, and U5N}

\item{nproc}{The number of processors to use}

\item{mutational_rate}{The rate of which light chains should mutate compared to heavy chains. 
Default is set to have light chain mutate at half the rate of heavy chains.}

\item{cell_id}{The name of the cell_id column in the naive BCR dataset(s)}

\item{sequence_alignment}{The name of the sequence_alinment column in the naive BCR dataset(s)}

\item{clone_id}{The name of the clone_id column in the naive BCR dataset(s)}

\item{sequence_id}{The name of the sequence_id column in the naive BCR dataset(s)}
}
\value{
A file of simulated sequences ready to be formatted into a \code{airrClone} 
object using \link{formatClones}
}
\description{
\code{getTreeSimulation} Simulates a dataset based on a given tree input within a \code{airrClone} objects.
}
