% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{hlClone}
\alias{hlClone}
\title{Example Multiple Partition Trees}
\format{
A data.frame with the following Change-O style columns:
  \itemize{
    \item  \code{clone_id}:           Clonal cluster
    \item  \code{data}:               List of airrClone objects
    \item  \code{locus}:              Locus identifier.
    \item  \code{seqs}:               Number of sequences
    \item  \code{parameters}:         Parameters associated with "raxml_trees"
    \item \code{trees}:               IgPhyML Tree
    \item \code{raxml_trees}:         RAxML Tree
}
}
\usage{
hlClone
}
\description{
A small example database originally created for the Immcantation package 'scoper.
}
\keyword{datasets}
