% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{imgtToIgblast}
\alias{imgtToIgblast}
\title{\code{imgtToIgblast} Format IMGT database to Igblast database}
\usage{
imgtToIgblast(
  imgt,
  outdir,
  igblast,
  organism = c("human", "mouse", "rhesus_monkey")
)
}
\arguments{
\item{imgt}{directory containing IMGT sequences ("dir" in downloadIMGT)}

\item{outdir}{output directory for IgBlast databases, will contain /fasta and /database}

\item{igblast}{IgBlast home directory location}

\item{organism}{organism species to use (must be human, mouse, or rhesus_monkey)}
}
\value{
outdir
}
\description{
Cleans IMGT database and creates blast dbs for IgBlast
}
\details{
Recoding of this script: 
https://bitbucket.org/kleinstein/immcantation/src/master/scripts/imgt2igblast.sh
}
