% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{installPythonDependencies}
\alias{installPythonDependencies}
\title{\link{installPythonDependencies} Checks for and installs the Python dependencies
                                 for UCA estimation}
\usage{
installPythonDependencies(python = "python3")
}
\arguments{
\item{python}{Specify the python call for your system. This is the call
on command line that issues the python you want to use.
"python3" by default.}
}
\description{
\link{installPythonDependencies} Checks for and installs the Python dependencies
                                 for UCA estimation
}
