% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{IsotypeTrees}
\alias{IsotypeTrees}
\title{Example Ig lineage trees with isotype reconstructions.}
\format{
A tibble of airrClone and phylo objects output by getTrees.
  \itemize{
    \item  \code{clone_id}:   Clonal cluster
    \item  \code{data}:       List of airrClone objects
    \item  \code{seqs}:       Number of sequences
    \item  \code{trees}:      List of phylo objects
  }
}
\usage{
IsotypeTrees
}
\description{
Same as ExampleClones but with isotypes predicted at internal nodes
}
\seealso{
\link{IsotypeTrees}
}
\keyword{datasets}
