% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTreesFunctions.R
\name{makeSkyline}
\alias{makeSkyline}
\title{get values for Bayesian Skyline plot}
\usage{
makeSkyline(
  object,
  bins = 100,
  youngest = 0,
  clone_id = NULL,
  max_height = c("min", "median", "mean", "max"),
  exclude_germline = TRUE
)
}
\arguments{
\item{object}{treedata object with parameters_posterior and trees_posterior}

\item{bins}{number of bins for plotting}

\item{youngest}{timepoint of the most recently tip sampled (if 0, backward time used)}

\item{clone_id}{name of the clone being analyzed (if desired)}

\item{max_height}{max height to use (min, median, mean, max)}

\item{exclude_germline}{exclude germline from skyline plot? (For TyCHE germline-rooted trees)}
}
\value{
Bayesian Skyline values for given clone
}
\description{
\code{makeSkyline}
}
