% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{mapSubtrees}
\alias{mapSubtrees}
\title{Get a map of nodes from one}
\usage{
mapSubtrees(treea, treeb)
}
\arguments{
\item{treea}{First phylo object}

\item{treeb}{Second phylo object}
}
\description{
Get a map of nodes from one
}
\details{
Returns a vector in which each position is the numerical
node number in treea, and the value is the corresponding node number in treeb
So node 1 in treea corresponds to result[1] in treeb
}
