% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{maskAmbigousSites}
\alias{maskAmbigousSites}
\title{\code{maskAmbigousSites} Masks ambiguous sites on V and J genes where references differ}
\usage{
maskAmbigousSites(
  clones,
  all_germlines,
  clone = "clone_id",
  mask_v = TRUE,
  mask_j = FALSE,
  mask_junc = FALSE,
  nproc = 1
)
}
\arguments{
\item{clones}{AIRR-table that is the output of \link{formatClones}.}

\item{all_germlines}{A data frame with all possible reconstructed germlines.}

\item{clone}{Column name for the clone ID.}

\item{mask_v}{Logical. Mask ambiguous germline sites in the V gene? Default is TRUE.}

\item{mask_j}{Logical. Mask ambiguous germline sites in the J gene? Default is FALSE.}

\item{mask_junc}{Logical. Mask ambiguous germline sites in the junction? Default is FALSE.}

\item{nproc}{Number of cores to use for parallel processing. Default is 1.}
}
\value{
An updated clones object with masked germlines.
}
\description{
This function masks ambiguous sites in the germline sequences of V and J genes
where references differ. It updates the clones object with masked germlines.
}
\examples{
\dontrun{
# Example usage:
clones <- maskAmbigousSites(clones, all_germlines)
}
}
