% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotMSA}
\alias{plotMSA}
\title{Simple function for plotting multiple sequence alignments
input can be either nucleotide or AA and contain ambiguous characters}
\usage{
plotMSA(
  data,
  id,
  seq,
  seq_type = c("nt", "aa"),
  text_size = 1.5,
  palette = NULL
)
}
\arguments{
\item{data}{Data table with sequences}

\item{id}{column name of sequence IDs (order will be preserved)}

\item{seq}{column name of sequence data to be plotted}

\item{seq_type}{sequence type in \code{seq} column, nucleotide (nt) or amino acid (aa)}

\item{text_size}{size of sequence alignment text}

\item{palette}{named vector of color for each character type in alignment, otherwise 
default nt or aa palette will be used}
}
\value{
a ggplot2-formatted multiple sequence alignment plot
}
\description{
Simple function for plotting multiple sequence alignments
input can be either nucleotide or AA and contain ambiguous characters
}
\seealso{
\link{getSeqPath}
}
