% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotSkylines}
\alias{plotSkylines}
\title{Simple function for plotting Bayesian skyline plots}
\usage{
plotSkylines(clones, file = NULL, width = 8.5, height = 11, ...)
}
\arguments{
\item{clones}{output from getTrees using BEAST}

\item{file}{pdf file name for printing plots}

\item{width}{width of plot in inches if file specified}

\item{height}{height of plot in inches if file specified}

\item{...}{optional arguments passed to grDevices::pdf}
}
\value{
if no file specified, a list of ggplot objects. If file specified
will plot to specified file
}
\description{
\code{plotSkylines} Simple Bayesian skyline plots
}
\seealso{
\link{getSkylines} \link{readBEAST} \link{getTrees}
}
