% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotTraces}
\alias{plotTraces}
\title{Simple function for plotting BEAST trace parameter traces}
\usage{
plotTraces(
  clones,
  burnin = 10,
  file = NULL,
  width = 8.5,
  height = 11,
  ess = TRUE,
  ...
)
}
\arguments{
\item{clones}{output from getTrees using BEAST}

\item{burnin}{percent of initial samples to discard (1-100)}

\item{file}{file name for printing plots}

\item{width}{width of plot in inches}

\item{height}{height of plot in inches}

\item{ess}{add ESS to facets?}

\item{...}{optional arguments passed to grDevices::pdf}
}
\value{
a PDF of trace plots
}
\description{
\code{plotTraces} exports trees to a pdf in an orderly fashion
}
\seealso{
\link{getTrees}
}
