% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readFasta}
\alias{readFasta}
\title{Read a fasta file into a list of sequences
\code{readFasta} reads a fasta file}
\usage{
readFasta(file)
}
\arguments{
\item{file}{FASTA file}
}
\value{
List of sequences
}
\description{
Read a fasta file into a list of sequences
\code{readFasta} reads a fasta file
}
