% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{readModelFile}
\alias{readModelFile}
\title{Read in a parsimony model file}
\usage{
readModelFile(file, useambig = FALSE)
}
\arguments{
\item{file}{parimony model file.}

\item{useambig}{use ambiguous naming as specified in the file?}
}
\value{
A named vector containing the states of the model
}
\description{
\code{readModelFile} Filler
}
\seealso{
\link{makeModelFile}, \link{bootstrapTrees}, \link{getTrees}
}
