% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clones.R
\name{readTreesAirr}
\alias{readTreesAirr}
\title{\code{readTreesJSON}
Experimental. Read trees from JSON/AIRR format from Dowser}
\usage{
readTreesAirr(file, heavy = "IGH")
}
\arguments{
\item{file}{.json file}

\item{heavy}{name of heavy chain locus}
}
\description{
\code{readTreesJSON}
Experimental. Read trees from JSON/AIRR format from Dowser
}
