% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{rootToTip}
\alias{rootToTip}
\title{Performs root-to-tip regression test on set of trees}
\usage{
rootToTip(
  trees,
  time = "time",
  permutations = 1000,
  germline = "Germline",
  minlength = 0.001,
  alternative = c("two.sided", "greater", "less")
)
}
\arguments{
\item{trees}{Tibble with trees and data}

\item{time}{Column id for timepoint}

\item{permutations}{Number of permutations for test}

\item{germline}{Germline sequence name}

\item{minlength}{Branch lengths to collapse in trees}

\item{alternative}{Perform one-sided (\code{greater} or \code{less})
or \code{two.sided} test}
}
\value{
A \code{tibble} with pearson correlation between divergene
and time, mean permuted correlation, p value(s), number of permutations,
and number of sequences
}
\description{
\code{rootToTop} performs root-to-tip regression permutation test
}
\details{
Output data table columns:
clone_id = clone id
observed = observed pearson correlation
permuted = mean permuted correlation
pgt = p value for DELTA < 0
plt = p value for DELTA > 0
}
\seealso{
Uses output from \link{getTrees}.
}
