% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clones.R
\name{sampleClone}
\alias{sampleClone}
\title{\code{sampleClone} Down-sample clone to specified size}
\usage{
sampleClone(clone, size, weight = NULL, group = NULL)
}
\arguments{
\item{clone}{an \link{airrClone} object}

\item{size}{target size}

\item{weight}{column for weighting sample probability}

\item{group}{column to sample evenly among groups}
}
\value{
a down-sampled airrClone object
}
\description{
\code{sampleClone} Down-sample clone to specified size
}
