% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clones.R
\name{sampleClones}
\alias{sampleClones}
\title{\code{sampleClones} Down-sample clones to specified size}
\usage{
sampleClones(clones, size, weight = NULL, group = NULL)
}
\arguments{
\item{clones}{a tibble of \link{airrClone} objects}

\item{size}{target size}

\item{weight}{column for weighting sample probability}

\item{group}{column (or columns) to sample evenly among groups}
}
\value{
The input object with sequences down-sampled
}
\description{
\code{sampleClones} Down-sample clones to specified size
}
