% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{scaleBranches}
\alias{scaleBranches}
\title{Scale branch lengths to represent either mutations or mutations per site.}
\usage{
scaleBranches(clones, edge_type = "mutations")
}
\arguments{
\item{clones}{a tibble of \code{airrClone} and \code{phylo} objects,
the output of \link{getTrees}.}

\item{edge_type}{Either \code{genetic_distance} (mutations per site) or 
\code{mutations}}
}
\value{
A tibble with \code{phylo} objects that have had branch lengths 
          rescaled as specified.
}
\description{
\code{scaleBranches} Branch length scaling function.
}
\details{
Uses clones$trees[[1]]$edge_type to determine how branches are currently scaled.
}
\seealso{
\link{getTrees}
}
