% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{setNodeDivergences}
\alias{setNodeDivergences}
\title{For each node in nodes list, store the divergence from the UCA}
\usage{
setNodeDivergences(tree)
}
\arguments{
\item{tree}{A \code{phylo} object with nodes list}
}
\value{
\code{phylo} with node list and divergences set
}
\description{
For each node in nodes list, store the divergence from the UCA
}
