% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clones.R
\name{stopCodonCheck}
\alias{stopCodonCheck}
\title{Check whether sequences have in-frame premature stop codons (PTCs)}
\usage{
stopCodonCheck(sequences, nproc = 1, check = TRUE)
}
\arguments{
\item{sequences}{Vector of nucleotide sequences in desired reading frame}

\item{nproc}{Number of cores for preprocessing}

\item{check}{Check whether codons split correctly}
}
\value{
Boolean vector of whether each sequence has an in-frame PTC
}
\description{
Check whether sequences have in-frame premature stop codons (PTCs)
}
