% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{testSP}
\alias{testSP}
\title{Performs SP (switch proportion) test on switch data}
\usage{
testSP(
  switches,
  permuteAll = FALSE,
  from = NULL,
  to = NULL,
  dropzeros = TRUE,
  bylineage = FALSE,
  pseudocount = 0
)
}
\arguments{
\item{switches}{Data frame from bootstrapTrees}

\item{permuteAll}{Permute among trees?}

\item{from}{Include only switches from this state?}

\item{to}{Include only switches to this state?}

\item{dropzeros}{Drop switches with zero counts?}

\item{bylineage}{Perform test for each lineage individually?}

\item{pseudocount}{Pseudocount for P value calculations}
}
\value{
A list containing a \code{tibble} with mean SP statistics, and another 
with SP statistics per repetition.
}
\description{
\code{testSP} performs an SP test
}
\details{
Output data table columns:
RECON = SP for observed data
PERMUTE = SP for permuted data
DELTA = RECON - PERMUTE
PLT = p value for DELTA < 0
PGT = p value for DELTA < 0
\itemize{
  \item  \code{FROM}: State going from.
  \item  \code{TO}: State going to.
  \item  \code{RECON}: SP for observed data.
  \item  \code{PERMUTE}: SP for permuted data.
  \item  \code{DELTA}:  RECON - PERMUTE.
  \item  \code{PLT}: p value that DELTA < 0
  \item  \code{PGT}: p value that DELTA > 0
  \item  \code{STAT}: Statistic used (SP).
  \item  \code{REP}: Bootstrap repetition.
  \item  \code{REPS}: Total number of ootstrap repetition.
}
}
\examples{
\dontrun{
igphyml <- "~/apps/igphyml/src/igphyml"
data(ExampleDb)
ExampleDb$sample_id = sample(ExampleDb$sample_id)
clones = formatClones(ExampleDb, trait="sample_id")
btrees = bootstrapTrees(clones[1:2], bootstraps=100, nproc=1,
   igphyml=igphyml, trait="sample_id", id="temp", dir="temp")
testSP(btrees$switches)
}
}
\seealso{
Uses output from \link{bootstrapTrees}. Related to \link{testPS}
and \link{testSC}.
}
