% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{TimeTrees}
\alias{TimeTrees}
\title{Example Ig lineage trees sampled over time.}
\format{
A tibble of airrClone and phylo objects output by getTrees.
  \itemize{
    \item  \code{clone_id}:   Clonal cluster
    \item  \code{data}:       List of airrClone objects
    \item  \code{seqs}:       Number of sequences
    \item  \code{trees}:      List of phylo objects
  }
}
\usage{
TimeTrees
}
\description{
Same as ExampleClones but with timepoint as a trait value
}
\seealso{
\link{TimeTrees}
}
\keyword{datasets}
