% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{updateAIRRGerm}
\alias{updateAIRRGerm}
\title{\code{updateAIRRGerm} Updates an AIRR-table to have a new germline alignment and associated germline statistics}
\usage{
updateAIRRGerm(
  airr_data,
  clones,
  igblast,
  igblast_database,
  references,
  organism,
  locus,
  outdir,
  nproc = 1
)
}
\arguments{
\item{airr_data}{the airr-table used to create the clones object}

\item{clones}{a clones object from \link{formatClones}}

\item{igblast}{the file path to igblast}

\item{igblast_database}{the file path to the igblast database}

\item{references}{the file path to the references to be used with igblast}

\item{organism}{the organism of the associated airr_data. Examples include "human" and "mouse"}

\item{locus}{if BCR data use "Ig" if TCR data use "TR"}

\item{outdir}{the directory to save intermediate files to}

\item{nproc}{the number of processors to use. Default is one.}
}
\value{
directory of downloaded IMGT BCR and TCR reference files
}
\description{
Uses the MRCA of a given clone as the germline alignment and updates the statistics by the associated IgBlast values
}
