% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{writeBeast2StrictXML}
\alias{writeBeast2StrictXML}
\title{Writes a BEAST XML file given an airrClone object
TODO: Make all model params optional}
\usage{
writeBeast2StrictXML(
  clone_data,
  time,
  dir = ".",
  include_gm = FALSE,
  mcmc_length = 1e+05,
  log_every = 10,
  germline = "Germline",
  units = "day",
  tree_prior = c("coalescent-skyline", "coalescent-constant"),
  pop_sizes = 5
)
}
\arguments{
\item{clone_data}{an \code{airrClone} object}

\item{time}{name of time column (trait in getTrees)}

\item{dir}{directory where temporary files will be placed}

\item{include_gm}{Include germline sequence?}

\item{mcmc_length}{Length of MCMC chain}

\item{germline}{name of germline sequence}

\item{units}{time units}

\item{tree_prior}{prior to use on tree}

\item{pop_sizes}{number of population sizes for coalescent-skyline prior}
}
\value{
The input clones tibble with an additional column for the bootstrap replicate trees.
}
\description{
\code{writeBeast2StrictXML} Phylogenetic bootstrap function.
}
