% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clones.R
\name{writeTreesAirr}
\alias{writeTreesAirr}
\title{\code{writeTreesAirr}
Experimental. Write trees in AIRR format}
\usage{
writeTreesAirr(
  object,
  file,
  repertoire_id = "sample",
  check = TRUE,
  verbose = TRUE,
  edge_tol = 1e-08,
  cell = "cell_id",
  heavy = "IGH",
  dowser_fields = TRUE
)
}
\arguments{
\item{object}{Dowser object from getTrees}

\item{file}{name of .json file}

\item{repertoire_id}{repertoire_id to use}

\item{check}{verify object is equivalent on reloading}

\item{verbose}{print out more info}

\item{edge_tol}{tolerance for branch length checks (if check=TRUE)}

\item{cell}{cell id column name in Dowser object}

\item{heavy}{name of heavy chain locus}

\item{dowser_fields}{include dowser-specific information? (recommended)}
}
\description{
\code{writeTreesAirr}
Experimental. Write trees in AIRR format
}
