% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{bootstrapTrees}
\alias{bootstrapTrees}
\title{Create a bootstrap distribution for clone sequence alignments, and estimate trees
 for each bootstrap replicate.}
\usage{
bootstrapTrees(
  data,
  bootstraps,
  nproc = 1,
  trait = NULL,
  dir = NULL,
  id = NULL,
  modelfile = NULL,
  dnapars = NULL,
  igphyml = NULL,
  trees = NULL,
  quiet = 0,
  rm_temp = FALSE,
  palette = NULL,
  resolve = 2,
  rep = NULL,
  keeptrees = TRUE,
  lfile = NULL
)
}
\arguments{
\item{data}{list of \code{changeoClone} objects}

\item{bootstraps}{number of bootstrap replicates to perform}

\item{nproc}{number of cores to parallelize computations}

\item{trait}{trait to use for parsimony models (required if \code{igphyml}
specified)}

\item{dir}{directory where temporary files will be placed (required
if \code{igphyml} or \code{dnapars} specified)}

\item{id}{unique identifer for this analysis (required if \code{igphyml}
or \code{dnapars} specified)}

\item{modelfile}{file specifying parsimony model to use}

\item{dnapars}{location of phylip dnapars executible if desired (not yet 
supported)}

\item{igphyml}{location of igphyml executible if trait models desired 
(optional)}

\item{trees}{tree topologies to use if aready available (bootstrapping
will not be perfomed)}

\item{quiet}{amount of rubbish to print to console}

\item{rm_temp}{remove temporary files (default=TRUE)}

\item{palette}{a named vector specifying colors for each state}

\item{resolve}{how should polytomies be resolved?}

\item{rep}{current bootstrap replicate (experimental)}

\item{keeptrees}{keep trees estimated from bootstrap replicates? (TRUE)}

\item{lfile}{lineage file input to igphyml if desired (experimental)}
}
\value{
A list of trees and/or switch counts for each bootstrap replicate.
}
\description{
\code{bootstrapTrees} Phylogenetic bootstrap function.
}
\details{
Tree building details are the same as \link{getTrees}. 
If \code{keeptrees=TRUE} (default) the returned object will contain a list named
 "trees" which contains a list of estimated tree objects for each bootstrap replicate.
 The object is structured like: trees[[<replicate>]][[<tree index>]]. If 
\code{igphyml} is specified (as well as \code{trait}, \code{temp}, and \code{id}),
 the returned object will contain a \code{tibble} named "switches" containing 
switch count information. This object can be passed to \link{testPS} and other 
functions to perform parsimony based trait value
tests.
}
\examples{
\dontrun{
data(ExampleDb)
ExampleDb$sample_id = sample(ExampleDb$sample_id)
clones = formatClones(ExampleDb, trait="sample_id")

btrees = bootstrapTrees(clones[1:2], bootstraps=100)
plotTrees(btrees$trees[[4]][[1]])

igphyml <- "~/apps/igphyml/src/igphyml"
btrees = bootstrapTrees(clones[1:2], bootstraps=100, nproc=1,
   igphyml=igphyml, trait="sample_id", id="temp", dir="temp")
plotTrees(btrees$trees[[4]][[1]])
testPS(btrees$switches)
}
}
\seealso{
Uses output from \link{formatClones} with similar arguments to 
\link{getTrees}. Output can be visualized with \link{plotTrees}, and 
tested with \link{testPS}, \link{testSC}, and \link{testSP}.
}
