% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{buildPML}
\alias{buildPML}
\title{Wrapper for phangorn::optim.pml}
\usage{
buildPML(
  clone,
  seq = "sequence",
  sub_model = "GTR",
  gamma = FALSE,
  asr = "seq",
  asr_thresh = 0.05,
  tree = NULL,
  data_type = "DNA",
  optNni = TRUE,
  optQ = TRUE,
  verbose = FALSE,
  resolve_random = TRUE
)
}
\arguments{
\item{clone}{\code{airrClone} object}

\item{seq}{sequece column in \code{airrClone} object}

\item{sub_model}{substitution model to use}

\item{gamma}{gamma site rate variation?}

\item{asr}{return sequence or probability matrix?}

\item{asr_thresh}{threshold for including a nucleotide as an alternative}

\item{tree}{fixed tree topology if desired.}

\item{data_type}{Are sequences DNA or AA?}

\item{optNni}{Optimize tree topology}

\item{optQ}{Optimize Q matrix}

\item{verbose}{Print error messages as they happen?}

\item{resolve_random}{randomly resolve polytomies?}
}
\value{
\code{phylo} object created by phangorn::optim.pml with nodes
         attribute containing reconstructed sequences.
}
\description{
Wrapper for phangorn::optim.pml
}
