% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{buildRAxML}
\alias{buildRAxML}
\title{Wrapper to build RAxML trees and infer intermediate nodes}
\usage{
buildRAxML(
  clone,
  seq = "sequence",
  exec,
  submodel = "GTRGAMMA",
  rseed = 28,
  n_searches = 1,
  name = "run",
  starting_tree = NULL,
  from_getTrees = FALSE,
  rm_files = TRUE,
  nproc = 1,
  asr = TRUE,
  partition = NULL,
  rep = 1,
  dir = NULL,
  ...
)
}
\arguments{
\item{clone}{list of \code{airrClone} objects}

\item{seq}{the phylo_seq option does this clone uses. Possible options are "sequence", "hlsequence", or "lsequence"}

\item{exec}{RAxML executable}

\item{submodel}{the model of binary (morphological), nucleotide, multi-state, or amino acid substitution 
Typical models are "GTRGAMMA", "GTRCAT", "GTRCATI", "ASC_GRTCAT", "ASC_GTRGAMMA", and "GTRGRAMMAI"}

\item{rseed}{The random seed used for the parsimony inferences. This allos you to reproduce your results.}

\item{n_searches}{specifies the number of alternative runs on distinct starting trees.}

\item{name}{specifies the name of the output file}

\item{starting_tree}{specifies a user starting tree file name in Newick format}

\item{from_getTrees}{A logical that indicates if the desired starting tree is from getTrees and not a newick file}

\item{rm_files}{remove temporary files?}

\item{nproc}{specifies the number of cores you want to run}

\item{asr}{computes the marginal ancestral states of a tree}

\item{partition}{specifies the file name of the assignment of models to 
alignment partitions for multiple models of substitution}

\item{rep}{Which repetition of the tree building is currently being run. Mainly for getBootstraps.}

\item{dir}{Where the output files are to be made.}

\item{...}{Additional arguments (not currently used)}
}
\value{
\code{phylo} object created by RAxML with nodes attribute
         containing reconstructed sequences.
}
\description{
Wrapper to build RAxML trees and infer intermediate nodes
}
