% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{calcRF}
\alias{calcRF}
\title{Finds the Robinson-Fould's Distance between phylogenies.}
\usage{
calcRF(tree_1, tree_2)
}
\arguments{
\item{tree_1}{A \code{phylo} object}

\item{tree_2}{A \code{phylo} object}
}
\value{
The input clones tibble with an additional column for the bootstrap replicate trees.
}
\description{
\code{calcRF} Calculates the RF distance between two phylogenetic trees with the same tips and tip labels.
}
