% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{createGermlines}
\alias{createGermlines}
\title{\link{createGermlines} Determine consensus clone sequence and create germline for clone}
\usage{
createGermlines(
  data,
  references,
  organism = "human",
  locus = "IGH",
  nproc = 1,
  seq_field = "sequence_alignment",
  v_field = "v_call",
  d_field = "d_call",
  j_field = "j_call",
  amino_acid = FALSE,
  id_field = "sequence_id",
  clone_field = "clone_id",
  j_germ_start = "j_germline_start",
  j_germ_end = "j_germline_end",
  j_germ_length = "j_germline_length",
  np1_length = "np1_length",
  np2_length = "np2_length",
  ...
)
}
\arguments{
\item{data}{AIRR-table containing sequences from one clone}

\item{references}{Full list of reference segments, see \link{readIMGT}}

\item{organism}{Species in \code{references} being analyzed}

\item{locus}{locus in \code{references} being analyzed}

\item{nproc}{Number of cores to use}

\item{seq_field}{Column name for sequence alignment}

\item{v_field}{Column name for V gene segment gene call}

\item{d_field}{Column name for D gene segment gene call}

\item{j_field}{Column name for J gene segment gene call}

\item{amino_acid}{Perform reconstruction on amino acid sequence (experimental)}

\item{id_field}{Column name for sequence ID}

\item{clone_field}{Column name for clone ID}

\item{j_germ_start}{Column name of index of J segment start within germline}

\item{j_germ_end}{Column name of index of J segment end within germline}

\item{j_germ_length}{Column name of index of J segment length within germline}

\item{np1_length}{Column name in receptor specifying np1 segment length}

\item{np2_length}{Column name in receptor specifying np2 segment length}

\item{...}{Additional arguments passed to \link{buildGermline}}
}
\value{
Tibble with reconstructed germlines
}
\description{
\link{createGermlines} Determine consensus clone sequence and create germline for clone
}
\details{
Return object adds/edits following columns:
\itemize{
  \item  \code{seq_field}:  Sequences potentially padded  same length as germline
  \item  \code{germline_alignment}: Full length germline
  \item  \code{germlien_alignment_d_mask}: Full length, D region masked
  \item  \code{vonly}:   V gene segment of germline if vonly=TRUE
  \item  \code{regions}: String of VDJ segment in position if useRegions=TRUE
}
}
\seealso{
\link{createGermlines} \link{buildGermline}, \link{stitchVDJ}
}
