% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clones.R
\name{createGermlines}
\alias{createGermlines}
\title{Wrapper for CreateGermlines.py}
\usage{
createGermlines(
  data,
  exec,
  refs,
  file,
  cf = "vj_clone",
  format = "airr",
  g = "dmask",
  germ = "germline_alignment_d_mask",
  rm_file = TRUE
)
}
\arguments{
\item{data}{tibble containing sequence information}

\item{exec}{location of CreateGermline.py}

\item{refs}{vector of reference allele locations}

\item{file}{temporary file name to write}

\item{cf}{column name for clone or subclone id}

\item{format}{airr or changeo format}

\item{g}{full or dmask germline option}

\item{germ}{name of the column containg germline DNA sequences in
output file.}

\item{rm_file}{remove temporary file?}
}
\value{
a tibble containing reconstructed germlines
}
\description{
\code{createGermlines} reconstructs clonal and subclonal germlines
}
