% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{dfToFasta}
\alias{dfToFasta}
\title{Write a fasta file of sequences
\code{readFasta} reads a fasta file}
\usage{
dfToFasta(
  df,
  file,
  id = "sequence_id",
  seq = "sequence",
  imgt_gaps = FALSE,
  columns = NULL
)
}
\arguments{
\item{df}{dataframe of sequences}

\item{file}{FASTA file for output}

\item{id}{Column name of sequence ids}

\item{seq}{Column name of sequences}

\item{imgt_gaps}{Keep IMGT gaps if present?}

\item{columns}{vector of column names to append to sequence id}
}
\value{
File of FASTA formatted sequences
}
\description{
Write a fasta file of sequences
\code{readFasta} reads a fasta file
}
