% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dowser.R
\docType{package}
\name{dowser-package}
\alias{dowser-package}
\title{dowser: B Cell Receptor Phylogenetics Toolkit}
\description{
Provides a set of functions for inferring, visualizing, and analyzing B cell phylogenetic trees. Provides methods to 1) reconstruct unmutated ancestral sequences, 2) build B cell phylogenetic trees using multiple methods, 3) visualize trees with metadata at the tips, 4) reconstruct intermediate sequences, 5) detect biased ancestor-descendant relationships among metadata types Workflow examples available at documentation site (see URL). Citations: Hoehn et al (2022) \doi{10.1371/journal.pcbi.1009885}, Hoehn et al (2021) \doi{10.1101/2021.01.06.425648}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dowser.readthedocs.io}
  \item Report bugs at \url{https://github.com/immcantation/dowser/issues}
}

}
\author{
\strong{Maintainer}: Kenneth Hoehn \email{kenneth.b.hoehn@dartmouth.edu}

Authors:
\itemize{
  \item Cole Jensen \email{cole.jensen@yale.edu}
  \item Jessie Fielding \email{jessie.jo.fielding@dartmouth.edu}
  \item Hunter Melton \email{hunter.j.melton@dartmouth.edu}
  \item Steven Kleinstein \email{steven.kleinstein@yale.edu} [copyright holder]
}

Other contributors:
\itemize{
  \item Susanna Marquez \email{susanna.marquez@yale.edu} [contributor]
  \item Jason Vander Heiden \email{jason.vanderheiden@gmail.com} [contributor]
}

}
\keyword{internal}
