% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dowser.R
\docType{package}
\name{dowser}
\alias{dowser}
\title{The dowser package}
\description{
\code{dowser} is a phylogenetic analysis package as part of the Immcantation suite of tools.
For additional details regarding the use of the \code{dowser} package see the 
vignettes:\cr
\code{browseVignettes("dowser")}
}
\references{
\enumerate{
  \item  Hoehn KB, Pybus OG, Kleinstein SH (2022) Phylogenetic analysis of 
        migration, differentiation, and class switching in B cells.
        PLoS Computational Biology. https://doi.org/10.1371/journal.pcbi.1009885
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dowser.readthedocs.io}
  \item Report bugs at \url{https://bitbucket.org/kleinstein/dowser/issues}
}

}
\author{
\strong{Maintainer}: Kenneth Hoehn \email{kenneth.b.hoehn@dartmouth.edu}

Authors:
\itemize{
  \item Steven Kleinstein \email{steven.kleinstein@yale.edu} [copyright holder]
}

Other contributors:
\itemize{
  \item Cole Jensen \email{cole.jensen@yale.edu} [contributor]
  \item Susanna Marquez \email{susanna.marquez@yale.edu} [contributor]
  \item Jason Vander Heiden \email{jason.vanderheiden@gmail.com} [contributor]
}

}
