% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dowser.R
\docType{package}
\name{dowser}
\alias{dowser}
\title{The dowser package}
\description{
\code{dowser} is a phylogenetic analysis package as part of the Immcantation suite of tools.
For additional details regarding the use of the \code{dowser} package see the 
vignettes:\cr
\code{browseVignettes("dowser")}
}
\references{
\enumerate{
  \item  Hoehn, KB, Pybus, OG, Kleinstein SH (submitted) Phylogenetic analysis of 
		migration, differentiation, and class switching	in B cells.
}
}
