% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{dowserObjectEquivalent}
\alias{dowserObjectEquivalent}
\title{\code{dowserObjectEquivalent}
Experimental. Check if two Dowser objects are equivalent}
\usage{
dowserObjectEquivalent(
  obj1,
  obj2,
  verbose = TRUE,
  edge_tol = 1e-06,
  dowser_fields = TRUE,
  nproc = 1
)
}
\arguments{
\item{obj1}{First Dowser object}

\item{obj2}{Second Dowser object}

\item{verbose}{print out more info}

\item{edge_tol}{tolerance for branch length checks (if check=TRUE)}

\item{dowser_fields}{check dowser-specific fields and gapped sequences?}

\item{nproc}{number of cores to use}
}
\description{
\code{dowserObjectEquivalent}
Experimental. Check if two Dowser objects are equivalent
}
\details{
In addition to the existing tree topology, edge length, sequence, and
data slot checks, this also verifies \code{tree$parameters} when present
-- including \code{build="pml"} trees.
}
