% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{exportTrees}
\alias{exportTrees}
\title{Exports the phylogenetic trees from the airrClone object}
\usage{
exportTrees(clones, filepath, tree_column = "trees", ...)
}
\arguments{
\item{clones}{tibble \code{airrClone} objects, the output of 
\link{getTrees} or \code{getTimeTrees}}

\item{filepath}{The file path for where the trees will be saved}

\item{tree_column}{The name of the column that contains the trees}

\item{...}{additional arguments to be passed}
}
\description{
\code{exportTrees}   Exports phylogenetic trees
}
