% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{exportTrees}
\alias{exportTrees}
\title{Exports the phylogentic trees from the airrClone object}
\usage{
exportTrees(clones, file, tree_column = "trees", ...)
}
\arguments{
\item{clones}{tibble \code{airrClone} objects, the output of 
\link{formatClones}}

\item{file}{The file path and name of where the trees will be saved}

\item{...}{additional arguments to be passed}
}
\description{
\code{exportTrees}   Exports phylogenetic trees
}
