% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTreesFunctions.R
\name{getDiffPoint}
\alias{getDiffPoint}
\title{Recurse up to tree to find most recent node with different state, or the root}
\usage{
getDiffPoint(
  tree,
  targetnode,
  trait,
  height = "height",
  verbose = FALSE,
  eo_adjust = FALSE,
  eo_type = NULL
)
}
\arguments{
\item{tree}{a treedata object, from getTimeTrees}

\item{targetnode}{current node}

\item{trait}{column name of the trait of interest}

\item{height}{which height value to return}

\item{verbose}{print out run info}

\item{eo_adjust}{adjust heights using expectOccupancies (recommended, requires eo_type)}

\item{eo_type}{if eo_adjust, trait value described by expectedOccupancies (typically state 1 of 2)}
}
\description{
\code{getDiffPoint}
}
