% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTreesFunctions.R
\name{getDiffPoints}
\alias{getDiffPoints}
\title{For each tree, recurse up to tree to find most recent 
node with a different state, or the root}
\usage{
getDiffPoints(
  data,
  trait,
  height = "CAheight_mean",
  verbose = FALSE,
  tip_traits = NULL
)
}
\arguments{
\item{data}{a tibble containing trees column from getTimeTrees}

\item{trait}{column name of the trait of interest}

\item{height}{which height value to return}

\item{verbose}{print out info during run}

\item{tip_traits}{vector of other traits to include for each tip.
Must have been also specified as traits in formatClones.}
}
\description{
\code{getDiffPoints}
}
\details{
Returns a data frame where each row is a tip in each tree
clone_id = clone id for that tree
tip = tip name
tip_tip = trait value for that tip
tip_height = height value for that tip
diff_node = most recent ancestor node with different trait value, or root
root = TRUE if at root node, FALSE otherwise
node_type = type of diff_node, will be "root" if at root node
node_height = height of diff_node 
<other columns> copied over from airrClone object for each tip
}
