% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{getDivergence}
\alias{getDivergence}
\title{Get divergence from root of tree for each tip}
\usage{
getDivergence(phy, minlength = 0.001)
}
\arguments{
\item{phy}{Tree object}

\item{minlength}{Branch lengths to collapse in trees}
}
\value{
A named vector of each tip's divergence from the tree's root.
}
\description{
\code{getDivergence} get sum of branch lengths leading from the 
root of the tree. If the germline sequence is included in the tree,
this will equal the germline divergence. If germline removed,
this will equal the MRCA divergence
}
