% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{getSkylines}
\alias{getSkylines}
\title{Make data frames for Bayesian skyline plots}
\usage{
getSkylines(
  clones,
  dir,
  time,
  bins = 100,
  verbose = 0,
  forward = TRUE,
  nproc = 1
)
}
\arguments{
\item{clones}{clone tibble}

\item{dir}{directory of BEAST trees file}

\item{time}{name of time column}

\item{bins}{number of bins for plotting}

\item{verbose}{if 1, print name of clones}

\item{forward}{plot in forward or (FALSE) backward time?}

\item{nproc}{processors for parallelization (by clone)}

\item{oldest}{age of the oldest tip sampled (if forward time desired)}
}
\value{
Bayesian Skyline values for given clone
}
\description{
\code{makeSkylines}
}
\details{
Burnin set from readBEAST or getTrees
}
