% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTreesFunctions.R
\name{getTimeTreesIterate}
\alias{getTimeTreesIterate}
\title{Iteratively resume getTimeTrees until convergence, as defined by 
all parameters (except those in \code{ignore} vector) having ESS 
greater than or equal to the specified ess_cutoff}
\usage{
getTimeTreesIterate(
  clones,
  template,
  beast,
  dir,
  id,
  time,
  iterations = 10,
  ess_cutoff = 200,
  burnin = 10,
  ignore = c("traitfrequencies"),
  continue = FALSE,
  posterior = c("none", "all", "parameters", "trees_with_traits", "trees"),
  trait = NULL,
  nproc = 1,
  quiet = 0,
  asr = FALSE,
  low_ram = TRUE,
  trim_ids = FALSE,
  ...
)
}
\arguments{
\item{clones}{a tibble of \code{airrClone} objects, the output of
\link{formatClones}}

\item{template}{XML template}

\item{beast}{location of beast binary directory (beast/bin)}

\item{dir}{directory where temporary files will be placed.}

\item{id}{unique identifer for this analysis}

\item{time}{Name of sample time column}

\item{iterations}{Maximum number of times to resume MCMC chain}

\item{ess_cutoff}{Minimum number of ESS for all parameters}

\item{burnin}{Burnin percent (default 10)}

\item{ignore}{Vector of parameters to ignore for ESS calculation}

\item{continue}{If TRUE, will check for iteration folder and resume from last iteration if found (default FALSE)}

\item{posterior}{Read un full distribution of parameters and trees? Can be "none" to just have
summary objects, "all" to have parameters, trees, and trees_with_traits, or a vector with the desired
combination of "parameters", "trees_with_traits", and "trees".}

\item{trait}{Trait column to be used}

\item{nproc}{Number of cores for parallelization. At most 1 core/tree can be used.}

\item{quiet}{quiet notifications if > 0}

\item{asr}{Log ancestral sequences?}

\item{low_ram}{run with less memory (slightly slower)}

\item{trim_ids}{remove last _ group from tips?}

\item{...}{Additional arguments for getTimeTrees}
}
\value{
A tibble of \code{tidytree} and \code{airrClone} objects.
}
\description{
\code{getTimeTreesIterate} Iteratively resume getTimeTrees til convergence.
}
\details{
For a full list of options (of which there are many more), see \code{getTimeTrees}

For examples and vignettes, see https://dowser.readthedocs.io
}
