% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{getTrees}
\alias{getTrees}
\title{Estimate lineage tree topologies, branch lengths,
and internal node states if desired}
\usage{
getTrees(
  data,
  trait = NULL,
  id = NULL,
  dir = NULL,
  modelfile = NULL,
  dnapars = NULL,
  igphyml = NULL,
  trees = NULL,
  nproc = 1,
  quiet = 0,
  rm_temp = FALSE,
  palette = NULL,
  resolve = 2
)
}
\arguments{
\item{data}{list of \code{changeoClone} objects}

\item{trait}{trait to use for parsimony models (required if 
\code{igphyml} specified)}

\item{id}{unique identifer for this analysis (required if 
\code{igphyml} or \code{dnapars} specified)}

\item{dir}{directory where temporary files will be placed (required 
if \code{igphyml} or \code{dnapars} specified)}

\item{modelfile}{file specifying parsimony model to use}

\item{dnapars}{location of phylip dnapars executible if desired (not yet supported)}

\item{igphyml}{location of igphyml executible if trait models desired (optional)}

\item{trees}{tree topologies to use if laready available}

\item{nproc}{number of cores to parallelize computations}

\item{quiet}{amount of rubbish to print to console}

\item{rm_temp}{remove temporary files (default=TRUE)}

\item{palette}{a named vector specifying colors for each state}

\item{resolve}{how should polytomies be resolved?}
}
\value{
A list of \code{phylo} objects in the same order as \code{data}.
}
\description{
\code{getTrees} Tree building function.
}
\details{
Estimates phylogenetic tree topologies and branch lengths for a list of 
\code{changeoClone} objects. By default, it will use phangnorn::pratchet to 
estimate maximum parsimony tree topologies, and  ape::acctran to estimate branch 
lengths. If \code{igpyhml} is specified, internal node \code{trait} values will 
be predicted by maximum parsimony. In this case, \code{dir} will need to be 
specified as a temporary directory to place all the intermediate files (will 
be created if not available). Further, \code{id} will need to specified to serve
 as a unique identifier for the temporary files. This should be chosen to ensure
 that multiple \code{getTrees} calls using the same \code{dir} do not overwrite 
each others files. 

\code{modelfile} is written automatically if not specified, but doesn't include
 any constraints. Intermediate files are deleted by default. This can be 
toggled using (\code{rm_temp}).
}
\examples{
\dontrun{
data(ExampleDb)
ExampleDb$sample_id = sample(ExampleDb$sample_id)
clones = formatClones(ExampleDb, trait="sample_id")

trees = getTrees(clones[1:2])
plotTrees(trees[[1]])

trees = getTrees(clones[1:2], igphyml=igphyml, id="temp", dir="temp", trait="sample_id")
plotTrees(trees[[1]])
}
}
\seealso{
\link{formatClones}, \link{bootstrapTrees}
}
