% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{getTreesAndUCAs}
\alias{getTreesAndUCAs}
\title{\link{getTreesAndUCAs} Construct trees and infer the UCA}
\usage{
getTreesAndUCAs(
  clones,
  data,
  dir = NULL,
  exec = NULL,
  resolve_germ = FALSE,
  repertoire_wide = FALSE,
  partition = "single",
  model_folder,
  model_folder_igk = NULL,
  model_folder_igl = NULL,
  python = "python3",
  id = "sample",
  max_iters = 100,
  nproc = 1,
  rm_temp = TRUE,
  quiet = 0,
  chain = "H",
  references = NULL,
  clone = "clone_id",
  cell = "cell_id",
  subsample_size = NA,
  search = "codon",
  check_genes = TRUE,
  fill_gaps = TRUE,
  genotyped = FALSE,
  split_light = FALSE,
  ...
)
}
\arguments{
\item{clones}{AIRR-table containing sequences \link{formatClones}}

\item{data}{The AIRR-table that was used to make the clones object.}

\item{dir}{The file path of the directory of where data is saved. NULL is default.}

\item{exec}{File path to the tree building executable}

\item{resolve_germ}{Resolve the V and J gene annotations within each clone?}

\item{repertoire_wide}{Build trees using parameters inferred from the entire dataset?}

\item{partition}{The partition model to use with IgPhyML. "single" is the default.}

\item{model_folder}{The file path to the OLGA default model files for heavy chains}

\item{model_folder_igk}{The file path to the OLGA default model files for IGK}

\item{model_folder_igl}{The file path to the OLGA default model files for IGL}

\item{python}{Specify the python call for your system. This is the call
on command line that issues the python you want to use. 
"python3" by default.}

\item{id}{The run ID, sample by default}

\item{max_iters}{The maximum number of iterations to run before ending. 
100 by default}

\item{nproc}{The number of cores to use}

\item{rm_temp}{Remove the generated files?}

\item{quiet}{Amount of noise to print out}

\item{chain}{Set to HL to use both heavy and light chain sequences}

\item{references}{Reference genes. See \link{readIMGT}}

\item{clone}{The name of the clone id column used in \link{formatClones}. 
If split_light was used in \link{formatClones}, use "clone_subgroup_id".}

\item{cell}{The name of the cell id in the AIRR table used to generate \link{formatClones}}

\item{subsample_size}{The amount that the clone should be sampled down to. Default is NA. Use NA if you do not wish to subsample -- in testing}

\item{search}{Search codon or nt space}

\item{check_genes}{Check if the inferred V/J lengths go into the inferred cdr3 region and adjust accordingly.}

\item{fill_gaps}{A logical that will fill in the V and J UCAs of clones that have partial V/J sequence alignments}

\item{genotyped}{A logical that indicates that AIRR-table has been genotyped and will only use the found annotations to create various germlines}

\item{split_light}{A logical that indicates if different light chain groups should be used to further split a clone (recommended for paired data)}

\item{...}{Additional arguments passed to various other functions like \link{getTrees} and \link{buildGermline}}
}
\value{
An \code{airrClone} object with trees and the inferred UCA
}
\description{
\link{getTreesAndUCAs} Construct trees and infer the UCA
}
\details{
Return object adds/edits following columns:
\itemize{
  \item  \code{trees}:  The phylogenies associated with each clone
  \item  \code{UCA}:    The inferred UCA
}
}
\seealso{
\link{getTrees}
}
