% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{getTreesAndUCAs}
\alias{getTreesAndUCAs}
\title{\link{getTreesAndUCAs} Construct trees and infer the UCA}
\usage{
getTreesAndUCAs(
  clones,
  data,
  dir = NULL,
  build,
  exec,
  model_folder,
  model_folder_igk = NULL,
  model_folder_igl = NULL,
  uca_script,
  python = "python",
  id = "sample",
  max_iters = 100,
  nproc = 1,
  rm_temp = TRUE,
  quiet = 0,
  chain = "H",
  omega = NULL,
  optimize = "lr",
  motifs = "FCH",
  hotness = "e,e,e,e,e,e",
  resolve_v = FALSE,
  resolve_j = FALSE,
  references = NULL,
  clone = "clone_id",
  heavy = "IGH",
  sampling_method = c("random", "lm", "ratio"),
  subsample_size = 5,
  method = "ML",
  rscript = NULL,
  check_genes = TRUE,
  ...
)
}
\arguments{
\item{clones}{AIRR-table containing sequences \link{formatClones}}

\item{data}{The AIRR-table that was used to make the clones object.}

\item{dir}{The file path of the directory of where data is saved. NULL is default.}

\item{build}{Name of the tree building method}

\item{exec}{File path to the tree building executable}

\item{model_folder}{The file path to the OLGA default model files for heavy chains}

\item{model_folder_igk}{The file path to the OLGA default model files for IGK}

\item{model_folder_igl}{The file path to the OLGA default model files for IGL}

\item{uca_script}{The file path to the UCA python script}

\item{python}{Specify the python call for your system. This is the call
on command line that issues the python you want to use. 
"python" by default.}

\item{id}{The run ID}

\item{max_iters}{The maximum number of iterations to run before ending}

\item{nproc}{The number of cores to use}

\item{rm_temp}{Remove the generated files?}

\item{quiet}{Amount of noise to print out}

\item{chain}{Set to HL to use both heavy and light chain sequences}

\item{omega}{Omega parameters to estimate (see IgPhyMl docs)}

\item{optimize}{Optimize HLP rates (r), lengths (l), and/or topology (r)}

\item{motifs}{Motifs to consider (see IgPhyMl docs)}

\item{hotness}{Hotness parameters to estimate (see IgPhyML docs)}

\item{resolve_v}{Resolve the V gene as well}

\item{resolve_j}{Resolve the J gene as well}

\item{references}{Reference genes. See \link{readIMGT}}

\item{clone}{The name of the clone id column used in \link{formatClones}}

\item{heavy}{The name of the heavy chain locus. Default is IGH.}

\item{sampling_method}{How to subsample. Methods include 'random', 'lm' (least mutated),
and 'ratio' (a weighted sampling with heavier weights
towards the least mutated sequences). The later two methods
require 'mu_freq' to be passed as a trait when running
\link{formatClones}}

\item{subsample_size}{The amount that the clone should be sampled down to. Default is 5.}

\item{method}{The method to find the UCA with. Options are ML (maximum likelihood), ML_tree, or MCMC.}

\item{rscript}{The file path to the rscript with get_UCA.py}

\item{check_genes}{Check if the inferred V/J lengths go into the inferred cdr3 region and adjust accordingly.}

\item{...}{Additional arguments passed to \link{buildGermline}}
}
\value{
An \code{airrClone} object with trees and the inferred UCA
}
\description{
\link{getTreesAndUCAs} Construct trees and infer the UCA
}
\details{
Return object adds/edits following columns:
\itemize{
  \item  \code{trees}:  The phylogenies associated with each clone
  \item  \code{UCA}:    The inferred UCA
}
}
\seealso{
\link{getTrees}
}
