% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{getTreesAndUCAs}
\alias{getTreesAndUCAs}
\title{\link{getTreesAndUCAs} Construct trees and infer the UCA}
\usage{
getTreesAndUCAs(
  clones,
  data,
  dir = NULL,
  build = "igphyml",
  exec = NULL,
  repertoire_wide = FALSE,
  partition = "single",
  model_folder,
  model_folder_igk = NULL,
  model_folder_igl = NULL,
  python = "python3",
  id = "sample",
  max_iters = 100,
  nproc = 1,
  rm_temp = TRUE,
  quiet = 0,
  chain = "H",
  references = NULL,
  clone = "clone_id",
  cell = "cell_id",
  sampling_method = "random",
  subsample_size = NA,
  search = "codon",
  check_genes = TRUE,
  igblast = NULL,
  igblast_database = NULL,
  ref_path = NULL,
  organism = "human",
  ...
)
}
\arguments{
\item{clones}{AIRR-table containing sequences \link{formatClones}}

\item{data}{The AIRR-table that was used to make the clones object.}

\item{dir}{The file path of the directory of where data is saved. NULL is default.}

\item{build}{Name of the tree building method. Currently only IgPhyML is supported.}

\item{exec}{File path to the tree building executable}

\item{repertoire_wide}{Build build trees using parameters inferred from the entire dataset?}

\item{partition}{The partition model to use with IgPhyML. "single" is the default.}

\item{model_folder}{The file path to the OLGA default model files for heavy chains}

\item{model_folder_igk}{The file path to the OLGA default model files for IGK}

\item{model_folder_igl}{The file path to the OLGA default model files for IGL}

\item{python}{Specify the python call for your system. This is the call
on command line that issues the python you want to use. 
"python3" by default.}

\item{id}{The run ID}

\item{max_iters}{The maximum number of iterations to run before ending}

\item{nproc}{The number of cores to use}

\item{rm_temp}{Remove the generated files?}

\item{quiet}{Amount of noise to print out}

\item{chain}{Set to HL to use both heavy and light chain sequences}

\item{references}{Reference genes. See \link{readIMGT}}

\item{clone}{The name of the clone id column used in \link{formatClones}}

\item{cell}{The name of the cell id in the AIRR table used to generate \link{formatClones}}

\item{sampling_method}{How to subsample. Methods include 'random', 'lm' (least mutated),
and 'ratio' (a weighted sampling with heavier weights
towards the least mutated sequences). The later two methods
require 'mu_freq' to be passed as a trait when running
\link{formatClones}}

\item{subsample_size}{The amount that the clone should be sampled down to. Default is NA. Use NA if you do not wish to subsample -- in testing}

\item{search}{Search codon or nt space}

\item{check_genes}{Check if the inferred V/J lengths go into the inferred cdr3 region and adjust accordingly.}

\item{igblast}{File path to igblast}

\item{igblast_database}{The file path to the database setup for igblast}

\item{ref_path}{The file path to your references parent folder.}

\item{organism}{The type of organism to align to if using igblast.}

\item{...}{Additional arguments passed to various other functions like \link{getTrees} and \link{buildGermline}}
}
\value{
An \code{airrClone} object with trees and the inferred UCA
}
\description{
\link{getTreesAndUCAs} Construct trees and infer the UCA
}
\details{
Return object adds/edits following columns:
\itemize{
  \item  \code{trees}:  The phylogenies associated with each clone
  \item  \code{UCA}:    The inferred UCA
}
}
\seealso{
\link{getTrees}
}
