% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{makeSkyline}
\alias{makeSkyline}
\title{get values for Bayesian Skyline plot}
\usage{
makeSkyline(
  logfile,
  treesfile,
  burnin,
  bins = 100,
  youngest = 0,
  clone_id = NULL
)
}
\arguments{
\item{logfile}{Beast log file}

\item{treesfile}{BEAST trees file}

\item{burnin}{file name for trace plots}

\item{bins}{number of bins for plotting}

\item{youngest}{timepoint of the most recently tip sampled (if 0, backward time used)}

\item{clone_id}{name of the clone being analyzed (if desired)}
}
\value{
Bayesian Skyline values for given clone
}
\description{
\code{makeSkyline}
}
