% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{readBEAST}
\alias{readBEAST}
\title{Reads in a BEAST output directory}
\usage{
readBEAST(
  clones,
  dir,
  exec,
  burnin = 10,
  nproc = 1,
  verbose = 1,
  tree_posterior = FALSE
)
}
\arguments{
\item{clones}{either a tibble (getTrees) or list of \code{airrClone} object}

\item{dir}{directory where temporary files will be placed}

\item{burnin}{percent of initial tree samples to discard (1-100)}

\item{nproc}{cores to use}

\item{verbose}{if > 0 print run information}

\item{tree_posterior}{Store tree posterior distribution? (used for density plot)}

\item{time}{name of time column (trait in getTrees)}
}
\value{
If data is a tibble, then the input clones tibble with additional columns for 
trees and parameter estimates given the specified burnin. If input is just a 
list of airrClone objects, it will return the corresponding list of trees
given the burnin
}
\description{
\code{readBEAST} Reads in data from BEAST output directory
}
