% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clones.R
\name{readTreesAirr}
\alias{readTreesAirr}
\title{\code{readTreesAirr}}
\usage{
readTreesAirr(file)
}
\arguments{
\item{file}{JSON file for output}
}
\value{
A Dowser object with airrClones and trees
trees = readTreesAirr("clones.json")
}
\description{
\code{readTreesAirr}
}
