% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readTreesJSON}
\alias{readTreesJSON}
\title{\code{readTreesJSON}
Experimental. Read trees from JSON/AIRR format from Dowser}
\usage{
readTreesJSON(
  file,
  heavy = "IGH",
  light = c("IGK", "IGL"),
  verbose = TRUE,
  edge_tol = 1e-08,
  nproc = 1
)
}
\arguments{
\item{file}{.json file}

\item{heavy}{name of heavy chain locus}

\item{light}{names of light chain loci}

\item{verbose}{how much info to print}
}
\description{
\code{readTreesJSON}
Experimental. Read trees from JSON/AIRR format from Dowser
}
\details{
Reads files written by \code{\link{writeTreesJSON}}, including trees built
with any \code{getTrees} \code{build} option.
}
