% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clones.R
\name{sampleCloneMultiGroup}
\alias{sampleCloneMultiGroup}
\title{\code{sampleCloneMultiGroup} Down-sample clone to specified size with multiple groups to sample evenly}
\usage{
sampleCloneMultiGroup(clone, size, weight = NULL, groups = NULL)
}
\arguments{
\item{clone}{an \link{airrClone} object}

\item{size}{target size}

\item{weight}{column for weighting sample probability}

\item{groups}{columns to sample evenly among groups}
}
\value{
a down-sampled airrClone object
}
\description{
\code{sampleCloneMultiGroup} Down-sample clone to specified size with multiple groups to sample evenly
}
