% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{treesEquivalent}
\alias{treesEquivalent}
\title{Check whether two tree objects are equivalent}
\usage{
treesEquivalent(
  obja,
  objb,
  edge_tol = 1e-08,
  numbering_match = FALSE,
  dowser_fields = TRUE
)
}
\arguments{
\item{obja}{First phylo or treedata object}

\item{objb}{Second phylo or treedata object}

\item{edge_tol}{tolerance for branch length checks (if check=TRUE)}

\item{numbering_match}{require internal node numbers to match?}

\item{verbose}{print out more info}
}
\description{
Check whether two tree objects are equivalent
}
\details{
For treedata objects, check both @phylo and @data
}
