% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFunctions.R
\name{writeCloneSequences}
\alias{writeCloneSequences}
\title{Write the sequences used in tree building to a fasta format. If there are more 
than one tree in airrClone output the sequence id will be followed by "|clone_id".}
\usage{
writeCloneSequences(clones, file)
}
\arguments{
\item{clones}{tibble \code{airrClone} objects, the output of 
\link{formatClones}}

\item{file}{The file path and name of where the sequences will be saved}
}
\description{
\code{writeCloneSequences}   Exports the sequences used in tree building.
}
