% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germlines.R
\name{writeFasta}
\alias{writeFasta}
\title{\code{writeFasta} Write a fasta file of sequences given a 
named list of sequences}
\usage{
writeFasta(seqs, file)
}
\arguments{
\item{seqs}{named list of sequences (output from \code{readFasta})}

\item{file}{FASTA file for output}
}
\value{
File of FASTA formatted sequences
}
\description{
\code{writeFasta} Write a fasta file of sequences given a 
named list of sequences
}
