% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clones.R
\name{writeTreesAirr}
\alias{writeTreesAirr}
\title{\code{writeTreesAirr}}
\usage{
writeTreesAirr(object, file, check = TRUE, verbose = TRUE, edge_tol = 1e-08)
}
\arguments{
\item{object}{Dowser tibble (output of formatClones or getTrees)}

\item{file}{JSON file for output}

\item{check}{Check if object saved faithfull}

\item{verbose}{Print number of checks if check=TRUE}

\item{edge_tol}{Allowable difference among edge lengths if check=TRUE
data(ExampleAirr)
# Select two clones, for demonstration purpose
sel <- c("3170", "3184")
clones <- formatClones(ExampleAirr[ExampleAirr$clone_id %in% sel,],traits="sample_id")
saveTreesAirr(clones, "clones.json")}
}
\description{
\code{writeTreesAirr}
}
